+++ podaac/subsetter/subset.py
@@ -24,7 +24,7 @@ import json
 import operator
 import os
 from itertools import zip_longest
-from typing import List, Tuple, Union
+from typing import List, Optional, Tuple, Union
 import dateutil
 from dateutil import parser
 
@@ -41,6 +41,8 @@ import xarray.coding.times
 from shapely.geometry import Point
 from shapely.ops import transform
 
+from podaac.subsetter import gpm_cleanup as gc
+from podaac.subsetter import time_converting as tc
 from podaac.subsetter import dimension_cleanup as dc
 from podaac.subsetter import xarray_enhancements as xre
 from podaac.subsetter.group_handling import GROUP_DELIM, transform_grouped_dataset, recombine_grouped_datasets, \
@@ -532,7 +534,7 @@ def compute_time_variable_name(dataset: xr.Dataset, lat_var: xr.Variable) -> str
         var_name_time = var_name.strip(GROUP_DELIM).split(GROUP_DELIM)[-1]
         if len(dataset[var_name].squeeze().dims) == 0:
             continue
-        if 'time' == var_name_time.lower() and dataset[var_name].squeeze().dims[0] in lat_var.squeeze().dims:
+        if ('time' == var_name_time.lower() or 'timeMidScan' == var_name_time) and dataset[var_name].squeeze().dims[0] in lat_var.squeeze().dims:
             return var_name
 
     # then check if any variables have 'time' in the string if the above loop doesn't return anything
@@ -742,7 +744,7 @@ def build_temporal_cond(min_time: str, max_time: str, dataset: xr.Dataset, time_
     return temporal_cond
 
 
-def get_base_group_names(lats):  # pylint: disable=too-many-branches
+def get_base_group_names(lats: List[str]) -> Tuple[List[str], List[Union[int, str]]]:  # pylint: disable=too-many-branches
     """Latitude groups may be at different depths. This function gets the level
     number that makes each latitude group unique from the other latitude names"""
     unique_groups = []
@@ -754,7 +756,7 @@ def get_base_group_names(lats):  # pylint: disable=too-many-branches
     # put the groups in the same levels in the same list
     group_list_transpose = np.array(group_list).T.tolist()
 
-    diff_count = ['' for i in range(len(group_list))]
+    diff_count = ['' for _ in range(len(group_list))]
     group_count = 0
     # loop through each group level
     for my_list in group_list_transpose:
@@ -786,7 +788,7 @@ def subset_with_bbox(dataset: xr.Dataset,  # pylint: disable=too-many-branches
                      lat_var_names: list,
                      lon_var_names: list,
                      time_var_names: list,
-                     variables=None,
+                     variables: Optional[List[str]] = None,
                      bbox: np.ndarray = None,
                      cut: bool = True,
                      min_time: str = None,
@@ -804,6 +806,8 @@ def subset_with_bbox(dataset: xr.Dataset,  # pylint: disable=too-many-branches
         Name of the longitude variables in the given dataset
     time_var_names : list
         Name of the time variables in the given dataset
+    variables : list[str]
+        List of variables to include in the result
     bbox : np.array
         Spatial bounding box to subset Dataset with.
     cut : bool
@@ -834,13 +838,6 @@ def subset_with_bbox(dataset: xr.Dataset,  # pylint: disable=too-many-branches
     else:
         unique_groups = [f'{GROUP_DELIM}{GROUP_DELIM.join(x.strip(GROUP_DELIM).split(GROUP_DELIM)[:-1])}' for x in lat_var_names]
 
-    # get unique group names for latitude coordinates
-    diff_count = [-1]
-    if len(lat_var_names) > 1:
-        unique_groups, diff_count = get_base_group_names(lat_var_names)
-    else:
-        unique_groups = [f'{GROUP_DELIM}{GROUP_DELIM.join(x.strip(GROUP_DELIM).split(GROUP_DELIM)[:-1])}' for x in lat_var_names]
-
     datasets = []
     total_list = []  # don't include repeated variables
     for lat_var_name, lon_var_name, time_var_name, diffs in zip(  # pylint: disable=too-many-nested-blocks
@@ -851,6 +848,11 @@ def subset_with_bbox(dataset: xr.Dataset,  # pylint: disable=too-many-branches
 
             if diffs == -1:  # if the lat name is in the root group: take only the root group vars
                 group_vars = list(dataset.data_vars.keys())
+                # include the coordinate variables if user asks for
+                group_vars.extend([
+                        var for var in list(dataset.coords.keys())
+                        if var in variables and var not in group_vars
+                    ])
             else:
                 group_vars = [
                     var for var in dataset.data_vars.keys()
@@ -859,13 +861,20 @@ def subset_with_bbox(dataset: xr.Dataset,  # pylint: disable=too-many-branches
                 # include variables that aren't in a latitude group
                 if variables:
                     group_vars.extend([
-                        var for var in dataset.data_vars.keys()
-                        if var in variables and var not in group_vars and var not in total_list and not var.startswith(tuple(unique_groups))
+                        var for var in dataset.variables.keys()
+                        if (var in variables and
+                            var not in group_vars and
+                            var not in total_list and
+                            not var.startswith(tuple(unique_groups))
+                            )
                     ])
                 else:
                     group_vars.extend([
                         var for var in dataset.data_vars.keys()
-                        if var not in group_vars and var not in total_list and not var.startswith(tuple(unique_groups))
+                        if (var not in group_vars and
+                            var not in total_list and
+                            not var.startswith(tuple(unique_groups))
+                            )
                         ])
 
                 # group dimensions do not get carried over if unused by data variables (MLS nTotalTimes var)
@@ -1020,64 +1029,24 @@ def get_coordinate_variable_names(dataset: xr.Dataset,
     return lat_var_names, lon_var_names, time_var_names
 
 
-def convert_to_datetime(dataset: xr.Dataset, time_vars: list) -> Tuple[xr.Dataset, datetime.datetime]:
-    """
-    Converts the time variable to datetime if xarray doesn't decode times
-
-    Parameters
-    ----------
-    dataset : xr.Dataset
-    time_vars : list
-
-    Returns
-    -------
-    xr.Dataset
-    datetime.datetime
-    """
-
-    for var in time_vars:
-        start_date = datetime.datetime.strptime("1993-01-01T00:00:00.00", "%Y-%m-%dT%H:%M:%S.%f")
-
-        if np.issubdtype(dataset[var].dtype, np.dtype(float)) or np.issubdtype(dataset[var].dtype, np.float32):
-            # adjust the time values from the start date
-            if start_date:
-                dataset[var].values = [start_date + datetime.timedelta(seconds=i) for i in dataset[var].values]
-                continue
-
-            utc_var_name = compute_utc_name(dataset)
-            if utc_var_name:
-                start_seconds = dataset[var].values[0]
-                dataset[var].values = [datetime.datetime(i[0], i[1], i[2], hour=i[3], minute=i[4], second=i[5]) for i in dataset[utc_var_name].values]
-                start_date = dataset[var].values[0] - np.timedelta64(int(start_seconds), 's')
-                return dataset, start_date
-
-        else:
-            pass
-
-    return dataset, start_date
-
-
 def open_as_nc_dataset(filepath: str) -> Tuple[nc.Dataset, bool]:
     """Open netcdf file, and flatten groups if they exist."""
-    file_extension = filepath.split('.')[-1]
+    hdf_type = None
+    # Open dataset with netCDF4 first, so we can get group info
+    try:
+        nc_dataset = nc.Dataset(filepath, mode='r')
+        has_groups = bool(nc_dataset.groups)
 
-    if file_extension == 'he5':
-        nc_dataset, has_groups = h5file_transform(filepath)
-    else:
-        # Open dataset with netCDF4 first, so we can get group info
-        try:
-            nc_dataset = nc.Dataset(filepath, mode='r')
-            has_groups = bool(nc_dataset.groups)
+        # If dataset has groups, transform to work with xarray
+        if has_groups:
+            nc_dataset = transform_grouped_dataset(nc_dataset, filepath)
 
-            # If dataset has groups, transform to work with xarray
-            if has_groups:
-                nc_dataset = transform_grouped_dataset(nc_dataset, filepath)
-        except OSError:
-            nc_dataset, has_groups = h5file_transform(filepath)
+    except OSError:
+        nc_dataset, has_groups, hdf_type = h5file_transform(filepath)
 
     nc_dataset = dc.remove_duplicate_dims(nc_dataset)
 
-    return nc_dataset, has_groups, file_extension
+    return nc_dataset, has_groups, hdf_type
 
 
 def override_decode_cf_datetime() -> None:
@@ -1161,7 +1130,8 @@ def subset(file_to_subset: str, bbox: np.ndarray, output_file: str,
         than one value in the case where there are multiple groups and
         different coordinate variables for each group.
     """
-    nc_dataset, has_groups, file_extension = open_as_nc_dataset(file_to_subset)
+    file_extension = os.path.splitext(file_to_subset)[1]
+    nc_dataset, has_groups, hdf_type = open_as_nc_dataset(file_to_subset)
 
     override_decode_cf_datetime()
 
@@ -1179,6 +1149,12 @@ def subset(file_to_subset: str, bbox: np.ndarray, output_file: str,
         lon_var_names = [var.replace('/', GROUP_DELIM) for var in lon_var_names]
         time_var_names = [var.replace('/', GROUP_DELIM) for var in time_var_names]
 
+    if '.HDF5' == file_extension:
+        # GPM files will have a timeMidScan time variable present
+        if '__FS__navigation__timeMidScan' in list(nc_dataset.variables.keys()):
+            gc.change_var_dims(nc_dataset, variables)
+            hdf_type = 'GPM'
+
     args = {
         'decode_coords': False,
         'mask_and_scale': False,
@@ -1193,19 +1169,33 @@ def subset(file_to_subset: str, bbox: np.ndarray, output_file: str,
 
     if min_time or max_time:
         args['decode_times'] = True
+        # check fill value and dtype, we know that this will cause an integer Overflow with xarray
+        if 'time' in nc_dataset.variables.keys():
+            try:
+                if nc_dataset['time'].getncattr('_FillValue') == nc.default_fillvals.get('f8') and \
+                 nc_dataset['time'].dtype == 'float64':
+                    args['mask_and_scale'] = True
+            except AttributeError:
+                pass
+
     with xr.open_dataset(
             xr.backends.NetCDF4DataStore(nc_dataset),
             **args
     ) as dataset:
+
+        original_dataset = dataset
+
         lat_var_names, lon_var_names, time_var_names = get_coordinate_variable_names(
             dataset=dataset,
             lat_var_names=lat_var_names,
             lon_var_names=lon_var_names,
             time_var_names=time_var_names
         )
+
         start_date = None
-        if file_extension == 'he5' and (min_time or max_time):
-            dataset, start_date = convert_to_datetime(dataset, time_var_names)
+        if hdf_type and (min_time or max_time):
+            dataset, start_date = tc.convert_to_datetime(dataset, time_var_names, hdf_type)
+
         chunks = calculate_chunks(dataset)
         if chunks:
             dataset = dataset.chunk(chunks)
@@ -1253,35 +1243,25 @@ def subset(file_to_subset: str, bbox: np.ndarray, output_file: str,
                     lon_var_names=lon_var_names
                 ))
             else:
-                encoding = {}
-                compression = {"zlib": True, "complevel": 5, "_FillValue": None}
-
-                if (min_time or max_time) and not all(
-                        dim_size == 1 for dim_size in dataset.dims.values()):
-                    encoding = {
-                        var_name: {
-                            'units': nc_dataset.variables[var_name].__dict__['units'],
-                            'zlib': True,
-                            "complevel": 5,
-                            "_FillValue": None
-                        } for var_name in time_var_names
-                        if 'units' in nc_dataset.variables[var_name].__dict__
-                    }
                 for var in dataset.data_vars:
-                    if var not in encoding:
-                        encoding[var] = compression
                     if dataset[var].dtype == 'S1' and isinstance(dataset[var].attrs.get('_FillValue'), bytes):
                         dataset[var].attrs['_FillValue'] = dataset[var].attrs['_FillValue'].decode('UTF-8')
 
+                    var_encoding = {
+                        "zlib": True,
+                        "complevel": 5,
+                        "_FillValue": original_dataset[var].encoding.get('_FillValue')
+                    }
+
                     data_var = dataset[var].copy()
-                    data_var.load().to_netcdf(output_file, 'a', encoding={var: encoding.get(var)})
+                    data_var.load().to_netcdf(output_file, 'a', encoding={var: var_encoding})
                     del data_var
 
                 with nc.Dataset(output_file, 'a') as dataset_attr:
                     dataset_attr.setncatts(dataset.attrs)
 
         if has_groups:
-            recombine_grouped_datasets(datasets, output_file, start_date)
+            recombine_grouped_datasets(datasets, output_file, start_date, time_var_names)
             # Check if the spatial bounds are all 'None'. This means the
             # subset result is empty.
             if any(bound is None for bound in spatial_bounds):